+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## [Unreleased]
 
+- Add option to remove "[]" from unpacked query openapi_parameters
+  ```ruby
+  OpenapiParameters::Query.new(
+    [parameter],
+    rack_array_compat: true
+  ).unpack('ids[]=2')
+  # => { 'ids' => ['2'] }
+  ```
+
 ## [0.3.4] - 2024-12-17
 
 - Add Parameter#convert(value)